homework9

See merge request nikitalisenkov/tfs_spring_2024!9

+++ app/src/main/java/com/example/app/data/network/model/TopicResponse.kt
@@ -9,7 +9,7 @@ data class TopicsResponse(
 
 data class TopicResponse(
     @SerializedName("max_id")
-    val maxId: Int,
+    val maxId: Long,
     @SerializedName("name")
     val name: String,
 )
